requirements.txt
autopep8==1.4.3
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0
djangorestframework==3.9.1
idna==2.8
mysqlclient==1.4.1
pep8==1.7.1
Pillow==5.4.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pytz==2018.9
requests==2.21.0
soupsieve==1.7.3
urllib3==1.24.1
